Ticket 48399 - Add helper makefile to lib389 to build and install
Bug Description:  It's not completely clear how to make a tarball or rpm from
the lib389 repo.

Fix Description:  This commit adds a helper make file and updates the readme
to support a build, install, test, rpm and srpm command. This will make it
easier to automate testing, and lower the barrier to entry to new developers.

https://fedorahosted.org/389/ticket/48399

Author: wibrown

Review by: vashirov