Ticket 47968 - Fix for format string in syslog.
Bug Description:  There is a format string error in log.c's usage of syslog (3).
This was detected in the RPM build which applies different settings.

Fix Description:  Fix the format string to be safe. Add the format string check
to the gcc-security check so that we can test for this.

https://fedorahosted.org/389/ticket/47968

Author: wibrown

Review by: One line rule.