Ticket 47944 - Implement sd_notify for systemd
Bug Description:  Systemd assumes processes are ready as soon as the PID file
is ready.

Fix Description:  By calling sd_notify, we are able to tell systemd when we are
actually ready to process requests (or as close to it as possible), and when
we are shutting down.

https://fedorahosted.org/389/ticket/47977

Author: wibrown

Review by: mreynolds (Thanks!)