Ticket #548 - RFE: Allow AD password sync to update shadowLastChange
Description: When passwordMinAge, passwordMaxAge, passwordWarning, etc.
are changed in a password policy, the corresponding shadow values are
also adjusted.

See this comment for more details.
https://fedorahosted.org/389/ticket/548#comment:19

This patch checks the current shadow values with the one from the
password policy, and if they don't match, it replaces the shadow value
with the one from the password policy.

Author: nhosoi

Review: wibrown