Ticket 47982 - improve timestamp resolution in logs
Bug Description:  An rfe was proposed to increase the resolution of timestamps
in logs to allow better automated machine sorting and time analysis of events.

Fix Description:  This patch enables a configuration option
nsslapd-logging-hr-timestamps-enabled, default to on that provides nanosecond
accurate timestamps in logs.

This is only avaliable on platforms that support clock_gettime.

This patch removes the str buffer as it actually slowed down logging performance.

This enables support in logconv.pl for HR timestamps.

https://fedorahosted.org/389/ticket/47982

Author: wibrown

Review by: nhosoi (Thanks!)