Ticket 47977 - Make systemd components depend on with-systemd
Bug Description:  The systemd unit files now depend on parts of the
dirsrv code base that depends on with-systemd (see, sd_notify)
As a result, it's possible to create a non-functional build of ds
by building without "with-systemd" and then attempting to run it.
The creates an issue where type=notify cannot be satisfied in the
start up procedure.

Fix Description: This makes all unit file components of systemd
depend on with-systemd in order to be processed and used.

https://fedorahosted.org/389/ticket/47977

Author: wibrown

Review by: nhosoi (Thanks)