Ticket 48431 - lib389 integrate ldclt
Bug Description:  ldclt is an object generation and load testing tool. It is
very useful for exposing issues in 389-ds when load is applied in different
ways.

Fix Description:  This adds initial integration of ldclt. This allows creation
of a set of users from template, and the test to bind to them rapidly. More
functions will be added as they are needed over time.

https://fedorahosted.org/389/ticket/48431

Author: wibrown

Review by: spichugi (Thank you!)