Ticket 48665 - Prevent sefault in ldbm_instance_modify_config_entry
Bug Description:  python ldap sends a mod_delete then a mod_add to the server.
This causes a segfault in ldbm_instance_modify_config_entry which expects the
value of the delete operation to exist.

Fix Description:  If there is no attribute in the delete operation, this now
passes a NULL to ldbm_config_set, which then rejects the op as NO_SUCH_ATTRIBUTE

There is an attached test case, which also checks that mod_replace is still
functional.

https://fedorahosted.org/389/ticket/48665

Author: wibrown

Review by: nhosoi (Thanks!)