Ticket 48342 - DNA: deadlock during DNA_EXTEND_EXOP_REQUEST_OID
Bug Description:  dna.c would deadlock during a range extension request.

This is because of lock ordering issues. In the normal operation, we would take:

* backend lock
* dna_lock

This is because *most* operations in dna are be_txn post operations.

However, when another replica requests a range, they would call the exop request
The issue with this is that the exop request is *not* a be_txn plugin. In fact
exop plugins were never able to have a be_txn type. So the code would take:

* dna_lock
* backend lock

This is how the dead lock starts. We have largely been lucky to not see this in
production before.

Fix Description:  This consumes the new RFE for betxn in plugin extendedop.
This means the locks are taken in the correct order, preventing the deadlock.

https://fedorahosted.org/389/ticket/48342

Author: wibrown

Review by: tbordaz and nhosoi (Thanks)