Ticket 389ds #48450 - Implement systemd password support for svrcore
Bug Description:  Svrcore is a secure password collection system. Is is used by
389 Directory Server to collect a pin from a text file, or if not avaliable,
it will ask the user. However, during system start up a tty is not always
avaliable to the user.

Fix Description:  This implement support for systemd's ask-password api as a
client. Additionally, we implement detection of tty connection to the user
module to help improve this. As a result, when a binary is run on the command
line, it will prompt using existing methods. If the tty is not avaliable, it
will contact systemd, which logs a wall message for the user to input a password

If the service is started from systemd, it will be integrated to the startup
command, IE:

$ systemctl start test-interactive.service
Enter PIN for internal (software): *****

https://fedorahosted.org/389/ticket/48450

Author: wibrown

Review by: nhosoi (Thanks!)