Ticket 48450 - Systemd password agent support
Bug Description:  Directory server needs to be able to prompt for passwords with
with systemd during systemd startup, or post start up.

Fix Description:  This allows Directory Server to take advantage of the svrcore
systemd integration, allowing systemd password as the "last resport" if file
or tty is not avaliable to the administrator.

https://fedorahosted.org/389/ticket/48450

Author: wibrown

Review by: nhosoi (Thanks!)