Ticket 48342 - Prevent transaction abort if a transaction has not begun
Bug Description:  Transactions may have been aborted if they had not begun yet
due to a logic issue in dna_update_config_event. Additionally, it was possible
for an operation to fail and the transaction to not be aborted, and for the
transaction to fail to start and the delete to proceed anyway!

Fix Description:  Re-arrange and correct the logic around the transaction
in dna_update_config_event. Given this code is always called during startup, we
do not have the same be-txn issues as other areas of dna. This should fix the
transaction logic.

https://fedorahosted.org/389/ticket/48342

Author: wibrown

Review by: tbordaz (Thanks)