Ticket 3 - headers contain ifdef blocks that should be removed.
Bug Description:  Using ifdef with_systemd in the header is not very clean
as it adds a dependency on clients to know about this.

Fix Description:  Move the ifdef into the systemd functions. The ifdef was
changed to HAVE_SYSTEMD. If not HAVE_SYSTEMD, functions return an error 10,
SVRCORE_MissingFeature, to indicate the feature was not compiled correctly.

https://pagure.io/svrcore/issue/3

Author: wibrown

Review by: nhosoi (Thanks!)