Ticket #48799 - objectclass values could be dropped on the consumer
Description: slapi_schema_expand_objectclasses was dropped if the modify
operation is replicated in the fix of Ticket 17 "new replication optimizations".

Since the replicated mods does not contain the expanded objectclasses,
slapi_schema_expand_objectclasses is needed on the consumer, too.

This patch resurrects the call for the consumer.

https://fedorahosted.org/389/ticket/48799

Author: nhosoi

Review: wibrown