Ticket #47536 - CI test: added test cases for ticket 47536
Description: Allow usage of OpenLDAP libraries that don't use NSS for crypto

OpenLDAP has an ability to tell the application which crypto library is
linked with.  The DS supports the both cases (OpenLDAP+NSS and OpenLDAP+
OpenSSL) by switching the behaviour based upon the info.  This test also
works for both.

This test also covers:
    Ticket #48756 - if startTLS is enabled, perl utilities fail to start.
    Config param nsslapd-extract-pemfiles

https://fedorahosted.org/389/ticket/48536

Reviewed by wibrown@redhat.com (Thank you, William!!!)