Ticket #48808 - Paged results search returns the blank list of entries
Bug Description: When a simple paged results slot in a connection is
discarded due to an error, e.g., SIZELIMIT_EXCEEDED, the slot was not
properly cleaned up.  Then, if the slot was reused, the leftover flag
confused the code and ended up returning the 0 search result.

Fix Description: This patch adds the clean up code when a slot is re-
used.

https://fedorahosted.org/389/ticket/48808

Reviewed by wibrown@redhat.com (Thank you, William!!)