Ticket 48434 - lib389 logging tools
Bug Description:  Many tests or error cases need to be able to check the log
of the system to determine if the issue has occured on not.

Fix Description:  This adds an initial construction of a log parser and
tools to interact with access and error logs so that we can use these effectively
in tests.

https://fedorahosted.org/389/ticket/48434

Author: wibrown

Review by: spichugi (Thanks!)