Ticket 48163 - Read schema from multiple locations
Bug Description:  We provide system schema that should not be altered or changed.
This schema tends to be version specific to ds versions also. As a result, it
would be good to move system schema to /usr, and keep custom or changing schema
in /etc.

Fix Description:  This adds a new configure option, systemschemadir, which
provides the location of system schema to the schema.c file. Additionally,
get priority filelist is altered to be able to prioritise over a number of
directories to form the filelist.

This leaves the only schema element in /etc as /etc/dirsrv/schema/99user.ldif

https://fedorahosted.org/389/ticket/48163

Author: wibrown

Review by: mreynolds (Thanks!)