Ticket #48822 - (389-ds-base-1.3.5) Fixing coverity issues.
Description: Buffer Overflow
BAD_SIZEOF -- slapd/ssl.c:3021: bad_sizeof: Taking the size of "randomPassword",
which is the address of an object, is suspicious.  Removed unnecessary cast.

SIZEOF_MISMATCH -- slapd/mapping_tree.c:2229: suspicious_sizeof: Passing
argument "errorbuf" of type "char *" and argument "8U /* sizeof (errorbuf) */"
to function "PL_strncpyz" is suspicious.

A pointer of an array is passed to a function, in which it tried to get the the
array size with sizeof, which does not work.  Changed the API to pass the size
of the array.

https://fedorahosted.org/389/ticket/48822

Reviewed by wibrown@redhat.com (Thank you, William!!)