Ticket 48818 - Fix case where return code is always -1
Bug Description:  Due to the docker process fixes, this made certain scripts
return the wrong code.

Fix Description:  Fix the location where we get the return code.

In addition, this fix showed an issue in the start-dirsrv script that has
existed for a long time. This corrects that issue also.

https://fedorahosted.org/389/ticket/48818

Author: wibrown

Review by: mreynolds (Thanks!)