Ticket 48617 - Server ram checks work in isolation
Bug Description:  Previously we would check all server cache allocations in
isolation. We would only know if I single cache would exceed our system ram.
This meant if you had many backends, it was possible to configure them each with
75% of the system ram, but be a valid configuration, even though it increased
the risk of oom.

Fix Description:  Now, all backends are check in isolation *and* they are check
as a whole regardless of manual or auto tuning. This also improves the auto
tuning system to check it's request, and to LOWER it if it is in excess of the
system ram. IE auto tuning a server should not ever be possible to create an OOM
condition.

https://fedorahosted.org/389/ticket/48617

Author: wibrown

Review by: nhosoi (Thanks!)