Ticket #48854 - Running db2index with no options breaks replication
Bug Description: It was a bug in import_foreman which was not adjusted
when the backend RUV entry was redesigned.  The backend RUV entry has
no parent entry by nature.  But import-forman mistakenly skipped to
handle the entry.  If the 'db2index' command line is executed on a
consumer, this error message is logged.

error log on a consumer:
 reindex userRoot: WARNING: Skipping entry "nsuniqueid=ffffffff-ffffffff-
 ffffffff-ffffffff" which has no parent, ending at line 18 of file "id2entry.db"
 reindex userRoot: WARNING: bad entry: ID 18

Due to this skip, the RUV entry is not found at the following start up
and a new RUV is generated by the MMR plugin. And the supplier finds the
generation ID mismatch.

error log on a supplier:
 NSMMReplicationPlugin - agmt="cn=0_1" (HOST:PORT): The remote replica
 has a different database generation ID than the local database.  You
 may have to reinitialize the remote replica, or the local replica.

Fix Description: Even if there is no parent entry, do not skip the RUV
entry as done for the suffix entry.

https://fedorahosted.org/389/ticket/48854

Reviewed by wibrown@redhat.com (Thank you, William!)

(cherry picked from commit ba3b8442abd4bd558d5467b07880d1c89613df9a)
(cherry picked from commit 14bc6ce2ba0e0613949dd45696b933e5a332a51c)