Ticket #48848 - modrdn deleteoldrdn can fail to find old attribute value, perhaps due to case folding
Bug Description: To replace an RDN witch type is case exact match fails
since ldbm_back_modrdn passed fully normalized value regardless of the
type.

Fix Description: Instead, it is being changed to pass the RDN maintaining
the original cases to moddn_newrdn_mods, in which the old rdn is correctly
identified both for the type ces and cis.

https://fedorahosted.org/389/ticket/48848