Ticket 48870 - Correct plugin execution order due to changes in exop
Bug Description:  Due to changes to make exop handling more efficient, an
issue was introduced where the execution of plugins was re-ordered. This meant
the precedence rules when two plugins are equal, were not the same.

Fix Description:  Return early when we find the plugin, rather than continuing
to loop to find the "last match"

https://fedorahosted.org/389/ticket/48870

Author: wibrown

Review by: nhosoi, tbordaz (Thanks, and sorry for breaking it ...)