Ticket 48877 - Fixes for RPM spec with spectool
Bug Description:  spectool expects a valid rpm file during parsing, but instead
it trims it after the %description ..... So instead, because we wrapped this
with an %if for asan, this now creates an incomplete and broken spec. This in
turn causes make rpms to fail in some cases.

Fix Description:  Move the if condition inside the %description, to prevent the
parser error.

https://fedorahosted.org/389/ticket/48877

Author: wibrown

Review by: nhosoi (Thanks!)