Ticket 48888 - Correction to create of dsldapobject
Bug Description:  We did not allow properly setting of DN for singletons in
create, or the many object case with the factor. This corrects that behaviour.

Fix Description:  If _dn is set, use it, else use the RDn + BaseDN to create
the dn of the object to be created

https://fedorahosted.org/389/ticket/48888

Author: wibrown

Review by: mreynolds (Thanks)