Ticket #48109 - substring index with nssubstrbegin: 1 is not being used with filters like (attr=x*)
Description: In case, index entry has this style of substr width definition:
   nsMatchingRule: nsSubstr{Begin,Middle,End}=<NUM>
it should be converted to
   nssubstr{Begin,Middle,End}: <NUM>
and skip the following nsMatchingRule evaluation.  There was a bug in the
logic to skip.  The feature itself was not effected, but this bogus error
was logged in the error log:
[..] from ldbm instance init: line 0: unknown or invalid matching rule
"nssubstrbegin=3" in index configuration (ignored)

Plus, the test script ticket48109_test.py is adjusted to the new format.

https://fedorahosted.org/389/ticket/48109

Reviewed by mreynolds@redhat.com (Thank you, Mark!!)

(cherry picked from commit 230ace2aa657e931b4c6dfa742a28d072a8f5db7)