Ticket #48892 - Wrong result code display in audit-failure log
Bug Description: Although a command line returns an error code 32,
audit-failure-log logs -1.

Fix Description: In the backend code, if a target entry does not
exist, -1 was internally set to ldap_result_code.  The code was
interpreted to LDAP_NO_SUCH_OBJECT in the frontend before returning
to the client.  But the audit-failure-log logged the internal code.
This patch fixes it.

https://fedorahosted.org/389/ticket/48892

Reviewed by wibrown@redhat.com (Thank you, William!!)