Ticket 48910 - Fixes for backend tests and lib389 reliability.
Bug Description:  During the update to lib389 a number of tests broke. Repeat
runs of the tests as a whole, if interrupted can cause failing of the next run.

This is because we re-use port 54321, bit with a different server id, so setup-ds
fails.

Fix Description:
* Fix the backend object type
* Change the serverid of all tests to be consistent.

https://fedorahosted.org/389/ticket/48910

Author: wibrown

Review by: spichugi (Thanks!)