Ticket 48916 - DNA Threshold set to 0 causes SIGFPE
Bug Description:  If the DNA threshold was set to 0, a divide by zero would
occur when requesting ranges.

Fix Description:  Prevent the config from setting a value of 0 for dna threshold.

If an existing site has a threshold of 0, we guard the divide operation, and
return an operations error instead.

https://fedorahosted.org/389/ticket/48916

Author: wibrown

Review by: nhosoi, mreynolds (Thank you!)