Ticket #48914 - db2bak.pl task enters infinitive loop when bak fs is almost full
Description: A backend helper function dblayer_copyfile returns an error
when any of the copy operation fails.  But one of the caller functions
dblayer_backup ignored the error.

This patch checks the error returned from dblayer_copyfile and abort the
back-up.

Also, more error info is added to the log messages.

https://fedorahosted.org/389/ticket/48914

Reviewed by mreynolds@redhat.com (Thank you, Mark!!)