Ticket 48925 - slapd crash with SIGILL: Dsktune should detect lack of CMPXCHG16B
Bug Description:  On older AMD the CMPXCHG16B is not present. This is critical
to the correct operation of lfds. Without out it we are unable to use nunc-stans

Fix Description:  dsktune should warn if CMPXCHG16B (flag cx16) is not present.
In a future release we will NOT allow installation upon a platform that lacks
this instruction.

https://fedorahosted.org/389/ticket/48925

Author: wibrown

Review by: nhosoi (Thank you!)