Ticket 48450 - Add prestart work around for systemd ask password
Bug Description:  Due to a lack of response to fix the systemd ask password
permissions, we must resolve this ourselves. Without this, we cannot utilise
the ask password feature at all.

Fix Description:  We add an execstartpre script, that parses dse.ldif for
the running server user. If found, we add the acl to ask-password directory
which will allow the server to start. We do this so that if each instance
has a unique user, they can all use ask pass correctly.

https://fedorahosted.org/389/ticket/48450

Author: wibrown

Review by: nhosoi (Thanks)