Ticket 397 - Add PBKDF2 to Directory Server password storage.
Bug Description:  We need to improve the cryptographic quality of hashes
available in DS to prevent attacks on hashes both online and offline.

Fix Description:  PBKDF2 is a hash that provides "iterations" of complexity
and work time to ensure complexity on the behalf of an attacker. It makes it
harder to create rainbow tables, bruteforce, or hardware accelerate attacks.

https://fedorahosted.org/389/ticket/397

Author: wibrown

Review by: rrelyea, nhosoi (Thanks!)