Ticket 48949 - os.makedirs() exist_ok not python2 compatible, added try/except
Bug Description:  os.makedirs on python3 takes an option that allows ignoring
the error if the directory does not exist in python2. This should be made more
portable by the use of the exception handling mechanism.

Fix Description:  Wrap the call in a try except to handle osError.

https://fedorahosted.org/389/ticket/48949

Author: xaellia

Review by: wibrown