Ticket 48949 - change default file path generation - use os.path.join
Bug Description:  Our path manipulation was done based on strings. This resulted
in a number of key locations being listed as //..../thing//file. This is not
very clean or professional, and may cause some issues with other tools.

Fix Description:  swap to using the python os.path.join mechanism which is able
to handle this correctly.

https://fedorahosted.org/389/ticket/48949

Author: xaellia

Review by: wibrown