Ticket 48949 - Fix ups for style and correctness
Bug Description:  We were provided with a method that set the name to "set"
rather than get. The method shouldn't be in the object, but as a standalone helper.
Additionally, os.path.join can take "path/dir" , "next/path", and join. We don't
need to seperate them up to a whole array.

Fix Description:  Fix the method to be get not set. change paths to have /
seperator when possible.

https://fedorahosted.org/389/ticket/48949

Author: wibrown

Review by: mreynolds (Thanks!)