Ticket 48951 - dsadm and dsconf base files
Fix Description:  To unite all commands within lib389! Add the base poc of dsadm
and dsconf. dsadm is able to list instances running on the server.

These are based on arg parse, and are extensible, able to be tested with pytest
and take advantage of all the features of lib389.

-- Move the SetupDs object out of tools to a seperate module in
the class hierarchy.

-- Complete rewrite of how the ini for the pre-rel installer works
this allows us to autogenerate examples, better manager merging options, and
track defaults. In the future we'll be able to base our versioned defaults
on this via subclassing.

-- So that we can assert the correct behaviour of the new installer
this allows us to unit test it, and assert the installation is correct as
we expect.

-- Add dsconf backend handling, and cli tests for the new
tools. This way we can make sure end to end of the process is tested, and that
the behaviour is correct as expected. This include option management and
failure scenarios we can't currently test in the cli tools.

-- Based on Mark's comments, add connOnly to the test and the cli.

https://fedorahosted.org/389/ticket/48951

Author: wibrown

Review by: nhosoi, mreynolds (Thanks!)