Ticket 48947 - Update default password hash to SSHA512
Bug Description:  Update the default password hash in DS to SSHA512

Fix Description:  This change fixes some of the macro placements, and makes
it easier to upgrade in the future. If we do the update, and the admin has not
set a hash in dse.ldif this will transparently upgrade the hashtype!

https://fedorahosted.org/389/ticket/48947

Author: wibrown

Review by: mreynolds, nhosoi (Thanks)