Ticket 48958 - Audit fail log doesn't work if audit log disabled.
Bug Description:  Due to a configuration interpretation issue, when audit was
not enabled, but auditfail was with no log defined, the fail log should write to
the audit log location on failed events, but audit events should not be written.
This did not work.

Fix Description:  This was because when we wrote to the audit file in the
abscence of the auditfail log, the audit enabled state was checked. This adds a
check to determine what the source event was from, and to check the correct log
enabled state during the event processing.

https://fedorahosted.org/389/ticket/48958

Author: wibrown

Review by: nhosoi (Thank you!)